feat: add fontfaceobserver dependency and implement resource loading logic
